# python_apps/workflow-engine-poc/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Unified Workflow Execution Engine",
        "description": "A clean, agnostic workflow execution engine PoC",
        "version": "1.0.0-poc"
    },
    "paths": {
...
        "/workflows/{workflow_id}/execute": {
            "post": {
                "tags": [
                    "workflows"
                ],
                "summary": "Execute Workflow By Id",
                "description": "Create a WorkflowExecution and kick off execution by workflow id.\n\nSupports both JSON and multipart/form-data payloads. For multipart, send:\n- payload: JSON string for non-file fields (trigger, user/session/org ids)\n- files[]: attachments for chat/file triggers",
                "operationId": "execute_workflow_by_id_workflows__workflow_id__execute_post",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Workflow Id"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_execute_workflow_by_id_workflows__workflow_id__execute_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowExecutionRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workflows/{workflow_id}/stream": {
            "get": {
                "tags": [
                    "workflows"
                ],
                "summary": "Stream Workflow Execution",
                "description": "Simple streaming endpoint that polls DB for the latest execution of the workflow and streams status updates.",
                "operationId": "stream_workflow_execution_workflows__workflow_id__stream_get",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Workflow Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workflows/validate": {
            "post": {
                "tags": [
                    "workflows"
                ],
                "summary": "Validate Workflow",
                "description": "Validate a workflow configuration",
                "operationId": "validate_workflow_workflows_validate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkflowBase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workflows/config": {
            "post": {
                "tags": [
                    "workflows"
                ],
                "summary": "Save Workflow",
                "description": "Save a workflow configuration to the database",
                "operationId": "save_workflow_workflows_config_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WorkflowBase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workflows/config/{workflow_id}": {
            "get": {
                "tags": [
                    "workflows"
                ],
                "summary": "Get Workflow",
                "description": "Get a workflow configuration by ID",
                "operationId": "get_workflow_workflows_config__workflow_id__get",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Workflow Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workflows/": {
            "get": {
                "tags": [
                    "workflows"
                ],
                "summary": "List Workflows",
                "description": "List all saved workflows returning ORM entities",
                "operationId": "list_workflows_workflows__get",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WorkflowRead"
                                    },
                                    "title": "Response List Workflows Workflows  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "workflows"
                ],
                "summary": "Create Workflow",
                "description": "Create a new workflow using SQLModel and return the SQLModel entity.",
                "operationId": "create_workflow_workflows__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": true,
                                "title": "Workflow Data"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/workflows/{workflow_id}": {
            "delete": {
                "tags": [
                    "workflows"
                ],
                "summary": "Delete Workflow",
                "description": "Delete a workflow configuration",
                "operationId": "delete_workflow_workflows__workflow_id__delete",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Workflow Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "workflows"
                ],
                "summary": "Get Workflow By Id",
                "description": "Get a workflow by ID using SQLModel and return the SQLModel entity.",
                "operationId": "get_workflow_by_id_workflows__workflow_id__get",
                "parameters": [
                    {
                        "name": "workflow_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Workflow Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WorkflowRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_execute_workflow_by_id_workflows__workflow_id__execute_post": {
                "properties": {
                    "payload": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Payload"
                    },
                    "files": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string",
                                    "format": "binary"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Files"
                    }
                },
                "type": "object",
                "title": "Body_execute_workflow_by_id_workflows__workflow_id__execute_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}